refactored profile/edit profile
